backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pandas
openpyxl
python-multipart
python-jose[cryptography]
passlib[bcrypt]
